Moved sbml tests into own file.
